+++ src/index.ts
@@ -1,2 +1,2 @@
-export { strapiLoader } from "./astro-loader.js";
-export { StrapiLoaderOptions } from "./options";
+export { strapiLoader } from "./strapi-loader";
+export type { StrapiLoaderOptions } from "./options";
